requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.41.0
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.0.1
numpy==2.3.2
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyngrok==7.3.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
ruff==0.12.7
safehttpx==0.1.6
safetensors==0.6.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
streamlit==1.48.0
tenacity==9.1.2
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
transformers==4.55.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
